requirements.txt
numpy==1.23.5
pandas==1.5.3
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
Pillow==8.3.2
requests==2.31.0  # Use `requests` instead of `wget`
DeepImageSearch
